feat: add album module with related migrations and routes
- Add album and album_files migrations to database schema
- Implement album CRUD operations and service layer
- Add album routes nested under members endpoint
- Create album models and response structures

fix: update task progress type from i32 to i64

- Change progress field in task messages table from i32 to i64
- Update progress handling in sync file task handler
- Fix progress clamping in file handler to use i64 values
- Rename SyncMessages entity to TaskMessages for clarity

refactor: reorganize thumbnail field in file_contents entity

- Move thumbnail field position in file_contents model
- Update entity relationships and imports accordingly

chore: enhance error handling with validation types

- Add Validation, NotFound, and Forbidden error types
- Update error categorization logic to handle new error types

+++ crates/schema/src/file.rs
@@ -84,7 +84,7 @@ pub struct TaskItemResponse {
     pub id: i64,
     pub task_type: String,
     pub status: String,
-    pub progress: i32,
+    pub progress: i64,
     pub message: String,
     pub created_at: String,           // 本地时区时间字符串
     pub updated_at: String,           // 本地时区时间字符串
